Create binapprox.py
In some situations, we do not need to compute the median exactly, and care more about
getting a quick approximation. Binapprox is a simple approximation algorithm derived from
binmedian. It follows the same steps as binmedian, except that it stops once it has computed
the median bin b, and just returns the midpoint of this interval. Hence to be perfectly clear,
the algorithm’s steps are:
1. Compute the mean µ and standard deviation σ
2. Form B bins across [µ − σ, µ + σ], map each data point to a bin
3. Find the bin b that contains the median
4. Return the midpoint of bin b
The median can differ from this midpoint by at most 1/2 the width of the interval, or
1/2 · 2σ/B = σ/B. Since we use B = 1000 in practice, binapprox is accurate to within
1/1000th of a standard deviation.